handling freed space; implemented index iterator

+++ DumpObject.h
@@ -16,13 +16,13 @@ class DumpObject
 protected:
     weak_ptr<WritableDump> dump;
     int64_t savedOffset;
-    int64_t savedLength;
+    int32_t savedLength;
 
     DumpObject(weak_ptr<WritableDump> dump);
     virtual void Write(ostream &stream) = 0;
 
 public:
     virtual void Write();
-    virtual int64_t NewLength() = 0;
+    virtual int32_t NewLength() = 0;
     int64_t SavedOffset();
 };
