+++ DumpTraits.h
@@ -21,7 +21,7 @@ public:
         value.Write(stream);
     }
 
-    static int64_t DumpSize()
+    static int32_t DumpSize()
     {
         return T::DumpSize();
     }
@@ -38,10 +38,10 @@ public:
         stream.read(bytes, 4);
 
         int32_t result = 0;
-        result |= (int64_t)bytes[0];
-        result |= (int64_t)bytes[1] << 8;
-        result |= (int64_t)bytes[2] << 16;
-        result |= (int64_t)bytes[3] << 24;
+        result |= (int32_t)(uint8_t)bytes[0];
+        result |= (int32_t)(uint8_t)bytes[1] << 8;
+        result |= (int32_t)(uint8_t)bytes[2] << 16;
+        result |= (int32_t)(uint8_t)bytes[3] << 24;
 
         return result;
     }
@@ -58,7 +58,7 @@ public:
         stream.write(bytes, 4);
     }
 
-    static int64_t DumpSize()
+    static int32_t DumpSize()
     {
         return 4;
     }
@@ -80,7 +80,7 @@ public:
         stream.write(&value, 1);
     }
 
-    static int64_t DumpSize()
+    static int32_t DumpSize()
     {
         return 1;
     }
